requirements.txt
coinmarketcap
pandas
python-binance
python-bittrex
sqlalchemy
mysql-connector==2.1.4